clean up java code.

+++ util/src/main/java/org/xipki/util/cbor/CborDecoder.java
@@ -777,7 +777,7 @@ public class CborDecoder implements AutoCloseable {
         }
 
         BigInteger[] ret = new BigInteger[arrayLen];
-        for (int i = 0; i < arrayLen.intValue(); i++) {
+        for (int i = 0; i < arrayLen; i++) {
             ret[i] = readBigInt();
         }
